+++ src/launcher/autoattach/wait_dll.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+#include <functional>
+
+namespace luadebug::autoattach {
+    bool wait_dll(bool(*loaded)(std::string const&));
+}
